+++ CHANGELOG.md
@@ -10,7 +10,7 @@ These changes are available in the [master branch](https://github.com/PrefectHQ/
 
 ### Enhancements
 
-- None
+- Don't create a `None` task for the `False` condition when using `ifelse` - [#2449](https://github.com/PrefectHQ/prefect/pull/2449)
 
 ### Task Library